requirements.txt
asgiref==3.2.3
autopep8==1.4.4
Django==3.0.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
Pillow==7.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
pytz==2019.3
sqlparse==0.3.0
